MAILBOX-315 Significantly improve CassandraMessageManagerTest speed
